requirements.txt
Django>=5.0,<6.0
tensorflow-intel==2.15.0
pandas>=2.0
numpy>=1.26
scikit-learn>=1.5
kagglehub
